refs #5 Add useI18next method for composition API
