+++ vue3-i18next/src/key.ts
@@ -1,0 +1,4 @@
+import { InjectionKey } from "vue";
+import { i18n } from "i18next";
+
+export const i18nextKey = Symbol("i18next") as InjectionKey<i18n>;
